Make template editor work at http://localhost:4200
Changed the way KeycloakUserHandler calculates the auth endpoint when
editor is run at http://localhost:4200. For now it takes the value
from window.authBase set in version.js by gulpfile.js. There's surely
a better way to do this.

For this to work  keycloak-realm.CEDAR.development.2023-07-05.json also
needs to be updated in cedar-docker-build:
{
    "clientId" : "cedar-angular-app",
    ...
    "redirectUris" : [ ..., "http://localhost:4200", "http://localhost:4200/*"],
    "webOrigins" : [ ..., "http://localhost:4200", "http://localhost:4200/*"],
 }